requirements.txt
Flask 
scikit-learn 
pyngrok
requests
python-dotenv
